Add files via upload
A stronger version of the usual brute force method of finding primes, using dynamic programming. This uses the fact that to prove a number is prime, it is sufficient to prove it has no prime factors- which means only prime divisors need to be checked. The script takes the input as an integer, and gives the number of primes under that number.